Upgraded to Angular v20, migrated away from angular cli ghpages, updated dependancies, and added X4 Blueprint Builder project.
Migrating Angular to v20 was mostly just swapping from *ngFor/*ngIf to @for and @if and removing useless settings that are now assumed.
Swapped to using normal GitHub actions for instead of angular-cli-ghpages with new commands for deploying and testing with act locally.
Removed or updated a bunch of dependencies that functionally changed nothing. Also updated SCSS which swapped some syntax for using variables/mixins from other files.